[tests] move fixtures to conftest.py
