+++ CONTRIBUTING.md
@@ -41,10 +41,10 @@ indexing and cleaning up sync events, and for particularly complex functions tha
 prone to regressions.
 
 The current test suite uses a Dropbox access token provided by the environment variable
-`DROPBOX_TOKEN` to connect to a real account.  The GitHub action which is running the
+`DROPBOX_TOKEN` to connect to a real account. The GitHub action which is running the
 tests will set this environment variable for you with a temporary access token that
-expires after 4 hours. Tests are run on `ubuntu- latest` and `macos- latest` in parallel
+expires after 4 hours. Tests are run on `ubuntu-latest` and `macos-latest` in parallel
 on different accounts and you should acquire a "lock" on the account before running
 tests. Fixtures to create and clean up a test config and to acquire a lock are provided
-in the `fixtures` module. If you run the tests locally, you will need to provide an
-access token for your own Dropbox account.
+in the `tests/linked/conftest.py`. If you run the tests locally, you will need to
+provide an access token for your own Dropbox account.
